requirements.txt
fastapi==0.121.0
uvicorn==0.38.0
paramiko==4.0.0
pydantic==2.12.4
python-dotenv==1.2.1
